Fixing build with clang

+++ src/Protocol/Protocol_1_8.cpp
@@ -1181,7 +1181,7 @@ void cProtocol_1_8_0::SendPlayerPermissionLevel()
 
 	cPacketizer Pkt(*this, pktEntityStatus);
 	Pkt.WriteBEUInt32(Player->GetUniqueID());
-	Pkt.WriteBEUInt8(Player->GetPermissionLevel() + 24);
+	Pkt.WriteBEUInt8(static_cast<UInt8>(Player->GetPermissionLevel() + 24));
 }
